requirements.txt
pytest
pytest-playwright
python-dotenv
playwright
pytest-bdd
pylint
black 